requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
ftfy
regex
tqdm
wandb
Pillow
git+https://github.com/openai/CLIP.git